requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==1.7.3
decorator==5.1.1
Django==4.2
django-browser-reload==1.13.0
django-crum==0.7.9
django-extensions==3.2.3
django-jazzmin==3.0.0
django-livereload-server==0.3.2
django-tailwind==3.1.0
django-widget-tweaks==1.4.8
exceptiongroup==1.2.1
executing==2.0.1
idna==3.7
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==2.0.0
opencv-python==4.10.0.84
parso==0.8.4
pillow==10.3.0
pip==24.1.1
poyo==0.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
plotly
pandas
django-select2